API Token authentification
The authentification for the API works with a token returned by the API after registering or authentificating. This token has to be sent with every request made by the client.
BaseController : this controller has to be extended by every controller of the API, so every response or error has the same format.
RegisterController : controller for the routes allowing a client to register or logging in.
Added routes for logging in and registering in the api.php file.